+++ config.yaml
@@ -54,9 +54,11 @@ ldap:
   url: ldap://abc.de
   user: user
   userdn: uid=%s,ou=people,dc=example,dc=de
+  useForLogin: true
 saml:
   idpMetadataURL: https://login.tum.de/idp/shibboleth
   idpName: TUM Login
+  idpColor: "#3070B3"
   cert: fullchain.pem
   privkey: privkey.pem
   entityID: https://live.rbg.tum.de/shib
